Moved all common functions to base controller
