+++ server/src/modules/role/index.js
@@ -3,9 +3,10 @@ const {
   authenticateRequest,
   authorizeRequest,
 } = require("../../common/middlewares");
+const { name: ModelName } = require("./model");
 
 const processRequest = async (req, res, next) => {
-  req.modelName = "Role";
+  req.modelName = ModelName;
   return next();
 };
